+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make it work with the officially released `rescript-relay` package.
 - Remove `vscode-graphiql-explorer` integration.
 - Add option to add query boilerplate for preloaded query.
+- Add step for selecting what type to expand when generating a query that uses the node interface/node top level field.
 
 ## 0.3.6